requirements.txt
Django==1.6
South==0.8.1
django-sortedm2m==0.6.1
django-model-utils==2.0.3
psycopg2==2.5.1
django-photologue==2.8
django-tastypie==0.9.16
django-redis-cache==0.13
django-registration==1.0
django-crispy-forms==1.4.0
django_mailgun==0.2.2
wikipedia==1.3.1
dj-database-url==0.3.0
dj-static==0.0.6
static==0.4
wsgiref==0.1.2
gunicorn==19.1.1
mimeparse==0.1.3
